split form to login and register
