+++ src/components/Form.vue
@@ -1,0 +1,14 @@
+<template>
+<div>
+  <div>
+    <router-view/>
+  </div>
+</div>
+</template>
+
+<script>
+</script>
+
+<style scoped>
+
+</style>
