+++ src/components/FormRegister.vue
@@ -1,0 +1,10 @@
+<template>
+<div>register form here</div>
+</template>
+
+<script>
+</script>
+
+<style scoped>
+
+</style>
