+++ src/router/index.js
@@ -9,7 +9,9 @@ import Homepage from '@/components/Homepage'
 import Dashboard from '@/components/Dashboard'
 import DashboardPrivate from '@/components/DashboardPrivate'
 import DashboardPublic from '@/components/DashboardPublic'
-import LoginForm from '@/components/LoginForm'
+import Form from '@/components/Form'
+import FormLogin from '@/components/FormLogin'
+import FormRegister from '@/components/FormRegister'
 
 Vue.use(Router)
 
@@ -67,8 +69,19 @@ export default new Router({
         },
         {
           path: 'login',
-          name: 'login',
-          component: LoginForm
+          component: Form,
+          children: [
+            {
+              path: '/login',
+              name: 'login',
+              component: FormLogin,
+            },
+            {
+              path: 'register',
+              name: 'register',
+              component: FormRegister,
+            }
+          ]
         }
       ]
     }
